Move all `get` responses into single directory
