+++ crates/sncast/src/response/get/class_hash_at.rs
