+++ crates/sncast/src/starknet_commands/get/block.rs
@@ -4,8 +4,8 @@ use sncast::get_block_id;
 use sncast::helpers::command::process_command_result;
 use sncast::helpers::configuration::CastConfig;
 use sncast::helpers::rpc::RpcArgs;
-use sncast::response::block::BlockResponse;
 use sncast::response::errors::{StarknetCommandError, handle_starknet_command_error};
+use sncast::response::get::block::BlockResponse;
 use sncast::response::ui::UI;
 use starknet_rust::providers::jsonrpc::HttpTransport;
 use starknet_rust::providers::{JsonRpcClient, Provider};
