+++ crates/sncast/src/starknet_commands/get/class_hash_at.rs
@@ -5,9 +5,9 @@ use sncast::get_block_id;
 use sncast::helpers::command::process_command_result;
 use sncast::helpers::configuration::CastConfig;
 use sncast::helpers::rpc::RpcArgs;
-use sncast::response::class_hash_at::ClassHashAtResponse;
 use sncast::response::errors::{StarknetCommandError, handle_starknet_command_error};
 use sncast::response::explorer_link::block_explorer_link_if_allowed;
+use sncast::response::get::class_hash_at::ClassHashAtResponse;
 use sncast::response::ui::UI;
 use starknet_rust::providers::jsonrpc::HttpTransport;
 use starknet_rust::providers::{JsonRpcClient, Provider};
