+++ crates/sncast/src/starknet_commands/get/tx_receipt.rs
@@ -4,7 +4,7 @@ use sncast::helpers::command::process_command_result;
 use sncast::helpers::configuration::CastConfig;
 use sncast::helpers::rpc::RpcArgs;
 use sncast::response::errors::{StarknetCommandError, handle_starknet_command_error};
-use sncast::response::tx_receipt::TxReceiptResponse;
+use sncast::response::get::tx_receipt::TxReceiptResponse;
 use sncast::response::ui::UI;
 use starknet_rust::providers::jsonrpc::HttpTransport;
 use starknet_rust::providers::{JsonRpcClient, Provider};
