+++ crates/sncast/src/starknet_commands/get/tx_status.rs
@@ -4,7 +4,9 @@ use sncast::helpers::command::process_command_result;
 use sncast::helpers::configuration::CastConfig;
 use sncast::helpers::rpc::RpcArgs;
 use sncast::response::errors::{StarknetCommandError, handle_starknet_command_error};
-use sncast::response::tx_status::{ExecutionStatus, FinalityStatus, TransactionStatusResponse};
+use sncast::response::get::tx_status::{
+    ExecutionStatus, FinalityStatus, TransactionStatusResponse,
+};
 use sncast::response::ui::UI;
 use starknet_rust::core::types::{TransactionExecutionStatus, TransactionStatus};
 use starknet_rust::providers::jsonrpc::HttpTransport;
